combined tables and fix custom query
