+++ src/Entity/CharCard.php
@@ -132,10 +132,34 @@ class CharCard
     protected $char_tier;
 
     /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\CharCardStat", inversedBy="char_card_stats")
-     * @ORM\JoinColumn(name="char_stat_id", referencedColumnName="id")
+     * @ORM\Column(type="integer")
      */
-    protected $char_stat_id;
+    protected $hit_points;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $attack;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $defense;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $agility;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $luck;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $speed;
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\Avatar", inversedBy="char_card_avatars")
@@ -237,6 +261,102 @@ class CharCard
     public function setCharStatId(CharCardStat $char_stat_id): void
     {
         $this->char_stat_id = $char_stat_id;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getHitPoints()
+    {
+        return $this->hit_points;
+    }
+
+    /**
+     * @param mixed $hit_points
+     */
+    public function setHitPoints($hit_points): void
+    {
+        $this->hit_points = $hit_points;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getAttack()
+    {
+        return $this->attack;
+    }
+
+    /**
+     * @param mixed $attack
+     */
+    public function setAttack($attack): void
+    {
+        $this->attack = $attack;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getDefense()
+    {
+        return $this->defense;
+    }
+
+    /**
+     * @param mixed $defense
+     */
+    public function setDefense($defense): void
+    {
+        $this->defense = $defense;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getAgility()
+    {
+        return $this->agility;
+    }
+
+    /**
+     * @param mixed $agility
+     */
+    public function setAgility($agility): void
+    {
+        $this->agility = $agility;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getLuck()
+    {
+        return $this->luck;
+    }
+
+    /**
+     * @param mixed $luck
+     */
+    public function setLuck($luck): void
+    {
+        $this->luck = $luck;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getSpeed()
+    {
+        return $this->speed;
+    }
+
+    /**
+     * @param mixed $speed
+     */
+    public function setSpeed($speed): void
+    {
+        $this->speed = $speed;
     }
 
     /**
