+++ src/Entity/CustomCard.php
@@ -28,15 +28,39 @@ class CustomCard
     protected $type;
 
     /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\CustomCardStat", inversedBy="custom_card_stats")
-     * @ORM\JoinColumn(name="card_stat_id", referencedColumnName="id")
+     * @ORM\Column(type="integer")
      */
-    protected $card_stat_id;
+    protected $votes;
 
     /**
      * @ORM\Column(type="integer")
      */
-    protected $votes;
+    protected $hit_points;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $attack;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $defense;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $agility;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $luck;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    protected $speed;
 
     /**
      * @return mixed
@@ -89,22 +113,6 @@ class CustomCard
     /**
      * @return mixed
      */
-    public function getCardStatId()
-    {
-        return $this->card_stat_id;
-    }
-
-    /**
-     * @param mixed $card_stat_id
-     */
-    public function setCardStatId(CustomCardStat $card_stat_id): void
-    {
-        $this->card_stat_id = $card_stat_id;
-    }
-
-    /**
-     * @return mixed
-     */
     public function getVotes()
     {
         return $this->votes;
@@ -116,5 +124,101 @@ class CustomCard
     public function setVotes($votes): void
     {
         $this->votes = $votes;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getHitPoints()
+    {
+        return $this->hit_points;
+    }
+
+    /**
+     * @param mixed $hit_points
+     */
+    public function setHitPoints($hit_points): void
+    {
+        $this->hit_points = $hit_points;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getAttack()
+    {
+        return $this->attack;
+    }
+
+    /**
+     * @param mixed $attack
+     */
+    public function setAttack($attack): void
+    {
+        $this->attack = $attack;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getDefense()
+    {
+        return $this->defense;
+    }
+
+    /**
+     * @param mixed $defense
+     */
+    public function setDefense($defense): void
+    {
+        $this->defense = $defense;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getAgility()
+    {
+        return $this->agility;
+    }
+
+    /**
+     * @param mixed $agility
+     */
+    public function setAgility($agility): void
+    {
+        $this->agility = $agility;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getLuck()
+    {
+        return $this->luck;
+    }
+
+    /**
+     * @param mixed $luck
+     */
+    public function setLuck($luck): void
+    {
+        $this->luck = $luck;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getSpeed()
+    {
+        return $this->speed;
+    }
+
+    /**
+     * @param mixed $speed
+     */
+    public function setSpeed($speed): void
+    {
+        $this->speed = $speed;
     }
 }
