--- src/Entity/CustomCardStat.php
@@ -1,180 +1,0 @@
-<?php
-
-namespace App\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
-
-/**
- * @ORM\Entity(repositoryClass="App\Repository\CustomCardStatRepository")
- */
-class CustomCardStat
-{
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    protected $id;
-
-    /**
-     * @ORM\OneToMany(targetEntity="App\Entity\CustomCard", mappedBy="card_stat_id")
-     */
-    protected $custom_card_stats;
-
-    public function __construct()
-    {
-        $this->custom_card_stats = new ArrayCollection();
-    }
-
-    /**
-     * @return Collection
-     */
-    public function getCustomCardStats()
-    {
-        return $this->custom_card_stats;
-    }
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $hit_points;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $attack;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $defense;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $agility;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $luck;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    protected $speed;
-
-    /**
-     * @return mixed
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @param mixed $id
-     */
-    public function setId($id): void
-    {
-        $this->id = $id;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getHitPoints()
-    {
-        return $this->hit_points;
-    }
-
-    /**
-     * @param mixed $hit_points
-     */
-    public function setHitPoints($hit_points): void
-    {
-        $this->hit_points = $hit_points;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getAttack()
-    {
-        return $this->attack;
-    }
-
-    /**
-     * @param mixed $attack
-     */
-    public function setAttack($attack): void
-    {
-        $this->attack = $attack;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getDefense()
-    {
-        return $this->defense;
-    }
-
-    /**
-     * @param mixed $defense
-     */
-    public function setDefense($defense): void
-    {
-        $this->defense = $defense;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getAgility()
-    {
-        return $this->agility;
-    }
-
-    /**
-     * @param mixed $agility
-     */
-    public function setAgility($agility): void
-    {
-        $this->agility = $agility;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getLuck()
-    {
-        return $this->luck;
-    }
-
-    /**
-     * @param mixed $luck
-     */
-    public function setLuck($luck): void
-    {
-        $this->luck = $luck;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getSpeed()
-    {
-        return $this->speed;
-    }
-
-    /**
-     * @param mixed $speed
-     */
-    public function setSpeed($speed): void
-    {
-        $this->speed = $speed;
-    }
-}
